create a data bridge for better grouping
