+++ EMS/common-bundle/src/Contracts/Bridge/Core/CoreBridgeInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EMS\CommonBundle\Contracts\Bridge\Core;
+
+interface CoreBridgeInterface
+{
+    /** @return array<mixed> */
+    public function versions(): array;
+
+    public function data(): CoreDataBridgeInterface;
+}
