+++ EMS/common-bundle/src/Contracts/Bridge/Core/CoreDataBridgeInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EMS\CommonBundle\Contracts\Bridge\Core;
+
+interface CoreDataBridgeInterface
+{
+    /** @param array<mixed> $rawData */
+    public function create(string $contentType, array $rawData = []): int;
+}
