--- EMS/common-bundle/src/Contracts/Bridge/CoreBridgeInterface.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EMS\CommonBundle\Contracts\Bridge;
-
-interface CoreBridgeInterface
-{
-    /** @return array<mixed> */
-    public function versions(): array;
-
-    /** @param array<mixed> $rawData */
-    public function documentCreate(string $contentType, array $rawData = []): int;
-}
