+++ EMS/common-bundle/src/Resources/config/twig.xml
@@ -11,7 +11,7 @@
             <tag name="twig.extension" />
         </service>
         <service id="ems.twig.extension.core_bridge" class="EMS\CommonBundle\Twig\CoreBridgeRuntime">
-            <argument type="service" id="EMS\CommonBundle\Contracts\Bridge\CoreBridgeInterface" />
+            <argument type="service" id="EMS\CommonBundle\Contracts\Bridge\Core\CoreBridgeInterface" />
             <tag name="twig.runtime" />
         </service>
         <service id="ems.twig.extension.http_client" class="EMS\CommonBundle\Twig\HttpClientRuntime">
